Fix the betty requirements

+++ 0x10-variadic_functions/3-print_all.c
@@ -1,24 +1,43 @@
 #include "variadic_functions.h"
 /**
- * print_all - Print anything, an int, float, char or string
- * @format: addres of the string that compares the format
+ * func_char - prints a char
+ * @list: the arguments of the list
  * Return: Nothing
  */
 void func_char(va_list list)
 {
- 	 printf("%c", (char)va_arg(list, int));
+	printf("%c", (char) va_arg(list, int));
 }
+/**
+ * func_int - prints a char
+ * @list: the arguments of the list
+ * Return: Nothing
+ */
+
 void func_int(va_list list)
 {
 	printf("%d", va_arg(list, int));
-} 
+}
+/**
+ * func_float - prints a char
+ * @list: the arguments of the list
+ * Return: Nothing
+ */
+
 void func_float(va_list list)
 {
-	printf("%f",(float) va_arg(list, double));
+	printf("%f", va_arg(list, double));
 }
+/**
+ * func_string - prints a char
+ * @list: the arguments of the list
+ * Return: Nothing
+ */
+
 void func_string(va_list list)
 {
 	char *string = va_arg(list, char *);
+
 	if (string == 0)
 	{
 		printf("(nil)");
@@ -27,14 +46,18 @@ void func_string(va_list list)
 	printf("%s", string);
 }
 
+/**
+ * print_all - Print anything, an int, float, char or string
+ * @format: addres of the string that compares the format
+ * Return: Nothing
+ */
 
 void print_all(const char * const format, ...)
 {
 	va_list list;
 	int i  = 0, j = 0;
-	char *ptr = ""; 
+	char *ptr = "";
 	op_t format_struct[] = {
-		
 		{"c", func_char},
 		{"i", func_int},
 		{"f", func_float},
@@ -43,10 +66,10 @@ void print_all(const char * const format, ...)
 
 	va_start(list, format);
 
-	while (format[i] != 0 )
+	while (format[i] != 0)
 	{
 		j = 0;
-		while (format_struct[j].op != NULL) 
+		while (format_struct[j].op != NULL)
 		{
 			if (*format_struct[j].op == format[i])
 			{
